+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `guard_min_version()` function to ensure a minimum version of Castor is used
 * Edited the duration of update check from `60 days` to `24 hours`
 * Add `wait_for_http_response()` function for a more generic response check
+* Add `wait_for_docker_container()` function to wait for a docker container to be ready
 * [BC Break] Remove `callable $responseChecker` parameter from `wait_for_http_status()`
 
 ## 0.11.1 (2024-01-11)